feat: introduces a 'dropzone' package
