+++ src/Umbraco.Web.UI.Client/src/packages/media/dropzone/components/dropzone.element.ts
@@ -1,6 +1,7 @@
 import { UmbDropzoneManager } from '../dropzone-manager.class.js';
 import { UmbDropzoneSubmittedEvent } from '../dropzone-submitted.event.js';
-import { UmbFileDropzoneItemStatus, type UmbUploadableItem } from '../types.js';
+import type { UmbUploadableItem } from '../types.js';
+import { UmbFileDropzoneItemStatus } from '../constants.js';
 import { css, customElement, html, ifDefined, property, state } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import type { UUIFileDropzoneElement, UUIFileDropzoneEvent } from '@umbraco-cms/backoffice/external/uui';
