+++ src/Umbraco.Web.UI.Client/src/packages/media/dropzone/components/index.ts
