+++ src/Umbraco.Web.UI.Client/src/packages/media/dropzone/dropzone-submitted.event.ts
