+++ src/Umbraco.Web.UI.Client/src/packages/media/dropzone/manifests.ts
@@ -1,0 +1,3 @@
+import { manifests as modalManifests } from './modals/manifests.js';
+
+export const manifests: Array<UmbExtensionManifest> = [...modalManifests];
