+++ src/Umbraco.Web.UI.Client/src/packages/media/dropzone/modals/manifests.ts
