+++ src/Umbraco.Web.UI.Client/src/packages/media/dropzone/types.ts
@@ -1,3 +1,4 @@
+import type { UmbFileDropzoneItemStatus } from './constants.js';
 import type { UUIFileFolder } from '@umbraco-cms/backoffice/external/uui';
 import type { UmbAllowedMediaTypeModel } from '@umbraco-cms/backoffice/media-type';
 import type { UmbTemporaryFileModel } from '@umbraco-cms/backoffice/temporary-file';
@@ -37,12 +38,4 @@ export interface UmbAllowedChildrenOfMediaType {
 export interface UmbFileDropzoneProgress {
 	total: number;
 	completed: number;
-}
-
-export enum UmbFileDropzoneItemStatus {
-	WAITING = 'waiting',
-	COMPLETE = 'complete',
-	NOT_ALLOWED = 'not allowed',
-	CANCELLED = 'cancelled',
-	ERROR = 'error',
 }
