+++ src/Umbraco.Web.UI.Client/src/packages/media/media/constants.ts
@@ -1,5 +1,4 @@
 export * from './collection/constants.js';
-export * from './dropzone/constants.js';
 export * from './entity-actions/constants.js';
 export * from './entity-bulk-actions/constants.js';
 export * from './menu/constants.js';
@@ -11,7 +10,6 @@ export * from './search/constants.js';
 export * from './tree/constants.js';
 export * from './url/constants.js';
 export * from './workspace/constants.js';
-
 export * from './paths.js';
 
 export { UMB_MEDIA_VARIANT_CONTEXT } from './property-dataset-context/media-property-dataset-context.token.js';
