--- src/Umbraco.Web.UI.Client/src/packages/media/media/dropzone/constants.ts
@@ -1,1 +1,0 @@
-export { UMB_DROPZONE_MEDIA_TYPE_PICKER_MODAL } from './modals/dropzone-media-type-picker/dropzone-media-type-picker-modal.token.js';
