--- src/Umbraco.Web.UI.Client/src/packages/media/media/dropzone/manifests.ts
@@ -1,1 +1,0 @@
-export * from './modals/manifests.js';
