+++ src/Umbraco.Web.UI.Client/src/packages/media/media/index.ts
@@ -1,6 +1,5 @@
 export * from './components/index.js';
 export * from './constants.js';
-export * from './dropzone/index.js';
 export * from './reference/index.js';
 export * from './repository/index.js';
 export * from './search/index.js';
