+++ src/Umbraco.Web.UI.Client/src/packages/media/media/manifests.ts
@@ -1,6 +1,5 @@
 import { manifests as auditLogManifests } from './audit-log/manifests.js';
 import { manifests as collectionManifests } from './collection/manifests.js';
-import { manifests as dropzoneManifests } from './dropzone/manifests.js';
 import { manifests as entityActionsManifests } from './entity-actions/manifests.js';
 import { manifests as entityBulkActionsManifests } from './entity-bulk-actions/manifests.js';
 import { manifests as fileUploadPreviewManifests } from './components/input-upload-field/manifests.js';
@@ -20,7 +19,6 @@ import { manifests as workspaceManifests } from './workspace/manifests.js';
 export const manifests: Array<UmbExtensionManifest> = [
 	...auditLogManifests,
 	...collectionManifests,
-	...dropzoneManifests,
 	...entityActionsManifests,
 	...entityBulkActionsManifests,
 	...fileUploadPreviewManifests,
