+++ src/Umbraco.Web.UI.Client/src/packages/media/media/types.ts
@@ -5,7 +5,6 @@ import type { UmbContentDetailModel, UmbElementValueModel } from '@umbraco-cms/b
 
 export type * from './audit-log/types.js';
 export type * from './collection/types.js';
-export type * from './dropzone/types.js';
 export type * from './modals/types.js';
 export type * from './recycle-bin/types.js';
 export type * from './repository/types.js';
