+++ src/Umbraco.Web.UI.Client/tsconfig.json
@@ -68,6 +68,7 @@ DON'T EDIT THIS FILE DIRECTLY. It is generated by /devops/tsconfig/index.js
 			"@umbraco-cms/backoffice/document-blueprint": ["./src/packages/documents/document-blueprints/index.ts"],
 			"@umbraco-cms/backoffice/document-type": ["./src/packages/documents/document-types/index.ts"],
 			"@umbraco-cms/backoffice/document": ["./src/packages/documents/documents/index.ts"],
+			"@umbraco-cms/backoffice/dropzone": ["./src/packages/media/dropzone/index.ts"],
 			"@umbraco-cms/backoffice/entity-action": ["./src/packages/core/entity-action/index.ts"],
 			"@umbraco-cms/backoffice/entity-bulk-action": ["./src/packages/core/entity-bulk-action/index.ts"],
 			"@umbraco-cms/backoffice/entity-create-option-action": [
